refactor: 表示部分をpagesから別コンポーネントに分離
